tooling/pyproject.toml
[project]
name = "tooling"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "langchain>=0.3.27",
    "langchain-openai>=0.3.32",
    "mcp[cli]>=1.13.1",
    "pandas>=2.3.2",
    "pandasql>=0.7.3",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
]
